refactor: WebSocket 복구 로직 WebSocketProvider로 통합 및 재연결 처리 개선
